Alter existing reassignment test case to fail with persisted model
When the model being reassigned is loaded fresh from the database, the
counter is changed twice, causing existing tests to fail.

It appears that on a fresh model, the @_after_create_counter_called
instance variable on the association is set and so the extra callback in
lib/active_record/associations/builder/belongs_to.rb never runs again.

On a persisted model loaded back from the DB, it's called as well as the
replace code in lib/active_record/associations/belongs_to_association.rb
which updates counters.